fix: extract styles to dry up
