+++ components/BookCatalog/CheckInAndOut/CheckOut.tsx
@@ -7,17 +7,7 @@ import { DatePicker } from '@mui/x-date-pickers';
 import axios from 'axios';
 import dayjs from 'dayjs';
 import React, { useCallback, useContext, useState } from 'react';
-
-const style = {
-  position: 'absolute' as const,
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: 350,
-  bgcolor: 'background.paper',
-  boxShadow: 24,
-  padding: '32px 32px 12px 32px'
-};
+import { buttonStyle, scanButtonStyle, style } from './styles';
 
 interface CheckoutPostBody {
   memberId: string;
@@ -112,37 +102,10 @@ const CheckOut = () => {
       }, 1000);
     }
   };
-  const scanButtonStyle = {
-    fontSize: '0.9rem',
-    mx: 0,
-    color: 'black',
-    display: 'block',
-    whiteSpace: 'nowrap',
-    backgroundColor: theme.palette.primary.main,
-    border: `1px solid ${theme.palette.primary.main}`,
-    padding: '0rem 1rem',
-    '&:hover': {
-      backgroundColor: theme.palette.primary.main
-    }
-  };
-
-  const checkInAndOutButtonStyle = {
-    my: 2,
-    fontSize: '1rem',
-    mx: 1,
-    color: 'black',
-    display: 'block',
-    whiteSpace: 'nowrap',
-    backgroundColor: theme.palette.primary.main,
-    border: `0.1rem solid ${theme.palette.primary.main}`,
-    '&:hover': {
-      backgroundColor: theme.palette.primary.main
-    }
-  };
 
   return (
     <div>
-      <Button onClick={handleOpen} sx={checkInAndOutButtonStyle}>
+      <Button onClick={handleOpen} sx={buttonStyle}>
         Check Out
       </Button>
       <Modal
@@ -231,10 +194,7 @@ const CheckOut = () => {
                 marginTop: '16px'
               }}
             >
-              <Button
-                onClick={handleCheckInOrOut}
-                sx={checkInAndOutButtonStyle}
-              >
+              <Button onClick={handleCheckInOrOut} sx={buttonStyle}>
                 {buttonText}
               </Button>
             </Box>
